# docs/swagger.yaml
basePath: /api
definitions:
  models.Loan:
    properties:
      loan_date:
        type: string
      resource:
        $ref: '#/definitions/models.Resource'
      resource_id:
        type: integer
      return_date:
        type: string
      user:
        $ref: '#/definitions/models.User'
      user_id:
        type: integer
    type: object
  models.Resource:
    properties:
      author:
        type: string
      id:
        type: integer
      is_available:
        type: boolean
      title:
        type: string
      type:
        $ref: '#/definitions/models.ResourceType'
    type: object
  models.ResourceType:
    enum:
    - book
    - game
    - film
    - autre
    type: string
    x-enum-varnames:
    - Book
    - Game
    - Film
    - Autre
  models.User:
    properties:
      email:
        type: string
      firstname:
        type: string
      id:
        type: integer
      lastname:
        type: string
      nationality:
        type: string
      phone:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is a sample API for managing a library system.
  title: Library Management API
  version: "1.0"
paths:
  /loans:
    get:
      description: Retrieve all loans
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Loan'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get all loans
      tags:
      - Loans
    post:
      consumes:
      - application/json
      description: Loan resources to users
      parameters:
      - description: List of loans
        in: body
        name: loans
        required: true
        schema:
          items:
            $ref: '#/definitions/models.Loan'
          type: array
      produces:
      - application/json
      responses:
        "201":
          description: Loan(s) created successfully
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Loan resources
      tags:
      - Loans
  /resources:
    get:
      consumes:
      - application/json
      description: Retrieve all resources
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Resource'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get all resources
      tags:
      - Resources
    post:
      consumes:
      - application/json
      description: Create a new resource with the provided details
      parameters:
      - description: Resource details
        in: body
        name: resource
        required: true
        schema:
          $ref: '#/definitions/models.Resource'
      produces:
      - application/json
      responses:
        "201":
          description: Resource created successfully
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Create a new resource
      tags:
      - Resources
  /resources/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a resource by its ID
      parameters:
      - description: Resource ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Resource deleted successfully
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Delete a resource
      tags:
      - Resources
    put:
      consumes:
      - application/json
      description: Update an existing resource with the provided details
      parameters:
      - description: Resource ID
        in: path
        name: id
        required: true
        type: string
      - description: Resource details
        in: body
        name: resource
        required: true
        schema:
          $ref: '#/definitions/models.Resource'
      produces:
      - application/json
      responses:
        "200":
          description: Resource updated successfully
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Update an existing resource
      tags:
      - Resources
  /restitute:
    post:
      consumes:
      - application/json
      description: Restitute resources from loans
      parameters:
      - description: List of loans to restitute
        in: body
        name: loans
        required: true
        schema:
          items:
            $ref: '#/definitions/models.Loan'
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: Resources restituted successfully
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Restitute resources
      tags:
      - Loans
  /users:
    get:
      consumes:
      - application/json
      description: Get a list of all users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: List all users
      tags:
      - Users
    post:
      consumes:
      - application/json
      description: Create a new user with the provided details
      parameters:
      - description: User details
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "201":
          description: User created successfully
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Create a new user
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: Update an existing user with the provided details
      parameters:
      - description: User details
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: User updated successfully
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Update an existing user
      tags:
      - Users
  /users/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a user by their ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User deleted successfully
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Delete a user
      tags:
      - Users
swagger: "2.0"